backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
httpx
python-dotenv
networkx
